requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# FastAPI and web
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Data processing
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# MLflow for tracking
mlflow>=2.7.0

# S3/MinIO
boto3>=1.28.0

# Airflow
apache-airflow>=2.7.0

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# Monitoring (optionnel)
prometheus-client>=0.18.0

# Testing (optionnel)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0

